cv_backend/requirements.txt
PIL
Django==2.2.6
django-cors-headers==3.1.1
djangorestframework==3.10.3
PyMySQL==0.9.3
